w3c-reports/12092014/mac-wk2/runned-tests/websockets/interfaces/WebSocket/close/004-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/' failed: WebSocket is closed before the connection is established.

PASS WebSockets: close() return value 
